Move start_tcpdump_processes from windows_monitoring to pretest
Move start_tcpdump_processes function and its make_tcpdump,
make_local_tcpdump imports from windows_monitoring.py to pretest.py.
Update conftest.py to import start_tcpdump_processes from pretest
instead of windows_monitoring. Also make SESSION_IS_CONTAINER_RUNNING
a local variable in pytest_runtestloop.
