+++ nat-lab/tests/conftest_helpers/windows_monitoring.py
@@ -3,43 +3,12 @@ import itertools
 import subprocess
 import threading
 import time
-from contextlib import AsyncExitStack
 from datetime import datetime
 from tests.utils.connection import ConnectionTag
-from tests.helpers_setup_checks import OPENWRT_VM_TAGS
 from tests.utils.connection.docker_connection import container_id
-from tests.utils.connection_util import new_connection_raw, is_running
+from tests.utils.connection_util import is_running
 from tests.utils.logger import setup_log
-from tests.utils.tcpdump import make_local_tcpdump, make_tcpdump
 from typing import List
-
-
-async def start_tcpdump_processes(
-    exit_stack: AsyncExitStack,
-):
-    connections = []
-    for gw_tag in ConnectionTag:
-        if gw_tag in OPENWRT_VM_TAGS:
-            continue
-        if "_GW" in gw_tag.name:
-            if not await is_running(gw_tag):
-                continue
-            connection = await exit_stack.enter_async_context(
-                new_connection_raw(gw_tag)
-            )
-            connections.append(connection)
-    for conn_tag in [
-        ConnectionTag.DOCKER_DNS_SERVER_1,
-        ConnectionTag.DOCKER_DNS_SERVER_2,
-    ]:
-        if await is_running(conn_tag):
-            connection = await exit_stack.enter_async_context(
-                new_connection_raw(conn_tag)
-            )
-            connections.append(connection)
-
-    await exit_stack.enter_async_context(make_tcpdump(connections, session=True))
-    await exit_stack.enter_async_context(make_local_tcpdump())
 
 
 async def start_windows_vms_resource_monitoring(
